+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - `M_detectIdent`
  - `M_choicePlain`, `M_choiceEmptyPlain`, `M_consPlain`, `M_nilPlain`
+ - `M_listMapInPlace`, `M_listMapInPlaceI`
 
 ### Changed